refactor: useGetUser 훅으로 분리
