+++ package.json
@@ -17,7 +17,6 @@
     "@mui/icons-material": "^5.16.0",
     "@mui/material": "^5.16.0",
     "@mui/material-nextjs": "^5.15.11",
-    "@prisma/client": "^5.16.1",
     "firebase": "^10.12.2",
     "next": "14.2.4",
     "react": "^18",
